+++ frontend/src/utils/polls/presidentielle2022.ts
